+++ cli/src/commands/file/show.rs
@@ -15,6 +15,7 @@
 use std::io::Write as _;
 
 use clap_complete::ArgValueCompleter;
+use itertools::Itertools as _;
 use jj_lib::backend::BackendResult;
 use jj_lib::conflicts::MaterializedTreeValue;
 use jj_lib::conflicts::materialize_merge_result;
@@ -22,7 +23,6 @@ use jj_lib::conflicts::materialize_tree_value;
 use jj_lib::file_util::copy_async_to_sync;
 use jj_lib::fileset::FilePattern;
 use jj_lib::fileset::FilesetExpression;
-use jj_lib::merge::MergedTreeValue;
 use jj_lib::repo::Repo as _;
 use jj_lib::repo_path::RepoPath;
 use pollster::FutureExt as _;
@@ -34,7 +34,9 @@ use crate::cli_util::WorkspaceCommandHelper;
 use crate::cli_util::print_unmatched_explicit_paths;
 use crate::command_error::CommandError;
 use crate::command_error::user_error;
+use crate::commit_templater::TreeEntry;
 use crate::complete;
+use crate::templater::TemplateRenderer;
 use crate::ui::Ui;
 
 /// Print contents of files in a revision
@@ -51,6 +53,23 @@ pub(crate) struct FileShowArgs {
         add = ArgValueCompleter::new(complete::revset_expression_all),
     )]
     revision: RevisionArg,
+
+    /// Render each file metadata using the given template
+    ///
+    /// All 0-argument methods of the [`TreeEntry` type] are available as
+    /// keywords in the template expression. See [`jj help -k templates`] for
+    /// more information.
+    ///
+    /// If not specified, this defaults to the `templates.file_show` setting.
+    ///
+    /// [`TreeEntry` type]:
+    ///     https://jj-vcs.github.io/jj/latest/templates/#treeentry-type
+    ///
+    /// [`jj help -k templates`]:
+    ///     https://jj-vcs.github.io/jj/latest/templates/
+    #[arg(long, short = 'T')]
+    template: Option<String>,
+
     /// Paths to print
     #[arg(
         required = true,
@@ -73,6 +92,16 @@ pub(crate) fn cmd_file_show(
     // TODO: No need to add special case for empty paths when switching to
     // parse_union_filesets(). paths = [] should be "none()" if supported.
     let fileset_expression = workspace_command.parse_file_patterns(ui, &args.paths)?;
+    let template = {
+        let language = workspace_command.commit_template_language();
+        let text = match &args.template {
+            Some(value) => value.to_owned(),
+            None => workspace_command.settings().get("templates.file_show")?,
+        };
+        workspace_command
+            .parse_template(ui, &language, &text)?
+            .labeled(["file_show"])
+    };
 
     // Try fast path for single file entry
     if let Some(path) = get_single_path(&fileset_expression) {
@@ -83,7 +112,11 @@ pub(crate) fn cmd_file_show(
         }
         if !value.is_tree() {
             ui.request_pager();
-            write_tree_entries(ui, &workspace_command, [(path, Ok(value))])?;
+            let entry = TreeEntry {
+                path: path.to_owned(),
+                value,
+            };
+            write_tree_entries(ui, &workspace_command, &template, [Ok(entry)])?;
             return Ok(());
         }
     }
@@ -93,7 +126,10 @@ pub(crate) fn cmd_file_show(
     write_tree_entries(
         ui,
         &workspace_command,
-        tree.entries_matching(matcher.as_ref()),
+        &template,
+        tree.entries_matching(matcher.as_ref())
+            .map(|(path, value)| Ok((path, value?)))
+            .map_ok(|(path, value)| TreeEntry { path, value }),
     )?;
     print_unmatched_explicit_paths(ui, &workspace_command, &fileset_expression, [&tree])?;
     Ok(())
@@ -111,19 +147,22 @@ fn get_single_path(expression: &FilesetExpression) -> Option<&RepoPath> {
     }
 }
 
-fn write_tree_entries<P: AsRef<RepoPath>>(
+fn write_tree_entries(
     ui: &Ui,
     workspace_command: &WorkspaceCommandHelper,
-    entries: impl IntoIterator<Item = (P, BackendResult<MergedTreeValue>)>,
+    template: &TemplateRenderer<TreeEntry>,
+    entries: impl IntoIterator<Item = BackendResult<TreeEntry>>,
 ) -> Result<(), CommandError> {
     let repo = workspace_command.repo();
-    for (path, result) in entries {
-        let value = result?;
-        let materialized = materialize_tree_value(repo.store(), path.as_ref(), value).block_on()?;
+    for entry in entries {
+        let entry = entry?;
+        template.format(&entry, ui.stdout_formatter().as_mut())?;
+        let materialized =
+            materialize_tree_value(repo.store(), &entry.path, entry.value).block_on()?;
         match materialized {
             MaterializedTreeValue::Absent => panic!("absent values should be excluded"),
             MaterializedTreeValue::AccessDenied(err) => {
-                let ui_path = workspace_command.format_file_path(path.as_ref());
+                let ui_path = workspace_command.format_file_path(&entry.path);
                 writeln!(
                     ui.warning_default(),
                     "Path '{ui_path}' exists but access is denied: {err}"
@@ -143,7 +182,7 @@ fn write_tree_entries<P: AsRef<RepoPath>>(
                 ui.stdout_formatter().write_all(id.describe().as_bytes())?;
             }
             MaterializedTreeValue::Symlink { .. } | MaterializedTreeValue::GitSubmodule(_) => {
-                let ui_path = workspace_command.format_file_path(path.as_ref());
+                let ui_path = workspace_command.format_file_path(&entry.path);
                 writeln!(
                     ui.warning_default(),
                     "Path '{ui_path}' exists but is not a file"